Revert "fix "undefined reference to \"main\"""
This reverts commit 9788a25ac32d6ee64ad3dab1ec879899498160d0.
