+++ docs/graph/graph-matching/general-match.md
@@ -219,7 +219,7 @@ author: H-J-Granger, accelsao, Ir1d, Early0v0, Henry-ZHR, HeliumOI, AntiLeaf
 
 ??? note "[UOJ #79. 一般图最大匹配](https://uoj.ac/problem/79)"
     ```cpp
-    --8<-- "docs/graph/graph-matching/code/general-match/general-match_1.hpp"
+    --8<-- "docs/graph/graph-matching/code/general-match/general-match_1.cpp"
     ```
 
 ## 基于高斯消元的一般图匹配算法
@@ -227,9 +227,9 @@ author: H-J-Granger, accelsao, Ir1d, Early0v0, Henry-ZHR, HeliumOI, AntiLeaf
 ??? tip "提示"
     在阅读以下内容前，你可能需要先阅读“线性代数”部分中关于矩阵的内容：
     
-    - [矩阵](../../math/linear-algebra/matrix.md)
-    - [行列式](../../math/linear-algebra/determinant.md)
-    - [高斯消元](../../math/linear-algebra/gauss.md)
+    - [矩阵](/docs/math/linear-algebra/matrix.md)
+    - [行列式](/docs/math/linear-algebra/determinant.md)
+    - [高斯消元](/docs/math/linear-algebra/gauss.md)
 
 这一部分将介绍一种基于高斯消元的一般图匹配算法。与传统的带花树算法相比，它的优势在于更易于理解与编写，同时便于解决“最大匹配中的必须点”等问题；缺点在于常数比较大，因为高斯消元的 $O(n^3)$ 基本是跑满的，而带花树一般跑不满。
